% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/pedigree.format.R
\name{founderIDs}
\alias{founderIDs}
\title{Extract a vector of founder individuals from a pedigree}
\usage{
founderIDs(pedigree, pedigree.type = "simple")
}
\arguments{
\item{pedigree.type}{Defaults to "simple", can also accept "plink" which
is equivalent for for first 5 to 6 columns of a PLINK .ped file.}

\item{ped}{Pedigree object in "simple" format (Three columns for ANIMAL,
MOTHER and FATHER) or in "plink" format (Five to Six columns for FAMILY,
ANIMAL, FATHER, MOTHER, SEX and Phenotype, where the phenotype column is
optional). The simple argument can recognise the order of parents if
they are named sensibly. Run simple.ped.name.rules() for an example.}
}
\description{
This function extracts a vector of founder IDs from a pedigree object.
A founder is an individual that does not have parents within the pedigree.
}

