% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/pedigree.format.R
\name{pedigree.format.genedrop}
\alias{pedigree.format.genedrop}
\title{Format Pedigree for genedrop analysis}
\usage{
pedigree.format.genedrop(ped)
}
\arguments{
\item{ped}{Pedigree object. Run simple.ped.name.rules() for an example.}
}
\description{
This function formats the pedigree for downstream analysis.
}

