% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/simperSNP-package.r
\docType{package}
\name{simperSNP}
\alias{simperSNP}
\alias{simperSNP-package}
\title{Simulates SNP Datasets For Simple And Complex Pedigrees With Recombination.}
\description{
simperSNP.
}
\keyword{package}

