% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/simulateGenotypes.R
\name{simulateGenos}
\alias{simulateGenos}
\title{Simulate recombination events in R}
\usage{
simulateGenos(ped, pedigree.type = "simple", cM = NULL, cM.male = NULL,
  cM.female = NULL, founder.mafs = NULL, chromosome.ids = NULL,
  snp.names = NULL, map.distances = NULL, error.rate = 1e-04,
  missing.rate = 0.001, xover.min.cM = NULL, xover.min.cM.male = NULL,
  xover.min.cM.female = NULL)
}
\arguments{
\item{ped}{data.frame of pedigree in "simple" format (Three columns for
ANIMAL, MOTHER and FATHER) or in "plink" format (Five to Six columns for
FAMILY, ANIMAL, FATHER, MOTHER, SEX and Phenotype, where the phenotype
column is optional). Is formatted within this function using
pedigree.format()}

\item{pedigree.type}{Character. "simple" or "plink"}

\item{cM}{Numeric vector of sex-averaged map distances in centiMorgans. Not
required if sex specific map distances are defined. First position should
be given as 0cM}

\item{cM.male}{Numeric vector of Male map distances in centiMorgans}

\item{cM.female}{Numeric vector of Female map distances in centiMorgans}

\item{founder.mafs}{Numeric vector of Minor allele frequencies for each
marker}

\item{chromosome.ids}{Vector of chromosome or linkage group IDs for each
marker}

\item{snp.names}{Optional character vector of SNP names}

\item{map.distances}{Optional vector of map distances in base pairs for
GenABEL object. Not used in the simulation.}

\item{error.rate}{numeric. Default is 1e-4. Future versions will allow vector.}

\item{missing.rate}{numeric. Default is 0.001. Future versions will allow vector.}

\item{xover.min.cM}{numeric. Minimum cM distance between two crossovers}

\item{xover.min.cM.male}{numeric. Minimum cM distance between two crossovers in males}

\item{xover.min.cM.female}{numeric. Minimum cM distance between two crossovers in females}
}
\description{
This function simulates genotypes within a pedigree, based on centiMorgan map
distances. Outputs genotypes as a GenABEL gwaa object.
}

